Integrate daterange into datasets API
PBENCH-1125

Remove the `GET /datasets/daterange` API in favor of a new query parameter, as
`GET /datasets?daterange`. This allows getting the actual date range for any
filtered selection of datasets.
