+++ lib/pbench/client/__init__.py
@@ -39,7 +39,6 @@ class API(Enum):
 
     DATASETS = "datasets"
     DATASETS_CONTENTS = "datasets_contents"
-    DATASETS_DATERANGE = "datasets_daterange"
     DATASETS_DETAIL = "datasets_detail"
     DATASETS_INVENTORY = "datasets_inventory"
     DATASETS_LIST = "datasets_list"
